Added ability for FileBoxApp to use database connector.
FileBox -
Made the database connector creator fall back to MSqlite if it cannot load the appsettings.json

Changed Table to include a Type instead of the name of the class.

Changed name from Program.cs to Startup.cs

FileBox.Shared -
Added Helpers folder

FileBoxApp -
Added DatabaseConnector dependency as well as SQLitePCLRaw.bundle_green which is used with Microsoft.Data.Sqlite otherwise the DatabaseConnector wouldn't work on android devices.

Added SQLitePCL initalise Batteries and Batteries_V2.

Added DatabaseConnector Server to app
