+++ FileBox.Shared/Filebox.Shared.csproj
@@ -12,4 +12,8 @@
 		</Reference>
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Helpers\" />
+	</ItemGroup>
+
 </Project>
